+++ R/sample.R
@@ -205,7 +205,7 @@ sampleStratMemory <- function(x, size, replace, lonlat, ext=NULL, weights=NULL, 
 
 
 
-sampleStratified <- function(x, size, replace=FALSE, as.df=TRUE, as.points=FALSE, values=TRUE, cells=TRUE, xy=FALSE, ext=NULL, warn=TRUE, exp=5, weights=NULL, exhaustive=FALSE) {
+sampleStratified <- function(x, size, replace=FALSE, as.df=TRUE, as.points=FALSE, values=TRUE, cells=TRUE, xy=FALSE, ext=NULL, warn=TRUE, exp=5, weights=NULL, exhaustive=FALSE, lonlat) {
 
 	if (nlyr(x) > 1) {
 		x <- x[[1]]
@@ -215,7 +215,6 @@ sampleStratified <- function(x, size, replace=FALSE, as.df=TRUE, as.points=FALSE
 		error("spatSample", "x has no values")
 	}
 
-	lonlat <- is.lonlat(x, perhaps=TRUE, warn=FALSE)
 
 	if ((blocks(x, n=4)$n == 1) || exhaustive) {
 		res <- sampleStratMemory(x, size, replace, lonlat, ext, weights, warn)
@@ -569,8 +568,11 @@ setMethod("spatSample", signature(x="SpatRaster"),
 
 		if (as.raster) return(sampleRaster(x, size, method, replace, ext, warn))
 
+		lonlat <- is.lonlat(x, perhaps=TRUE, warn=FALSE)
+		if (lonlat) exact <- FALSE
+
 		if (method == "stratified") {
-			return( sampleStratified(x, size, replace=replace, as.df=as.df, as.points=as.points, cells=cells, values=values, xy=xy, ext=ext, warn=warn, exp=exp, weights=weights, exhaustive=exhaustive) )
+			return( sampleStratified(x, size, replace=replace, as.df=as.df, as.points=as.points, cells=cells, values=values, xy=xy, ext=ext, warn=warn, exp=exp, weights=weights, exhaustive=exhaustive, lonlat=lonlat) )
 		} else if (!is.null(weights)) {
 			error("spatSample", "argument weights is only used when method='stratified'")
 		}