+++ NEWS.md
@@ -4,7 +4,7 @@
 
 ## enhancements
 
-- `spatSample<SpatRaster>` and `spatSample<SpatExtent>` argument gain argument "exact=FALSE" to request the exact (but perhaps less regular) sample size for a regular sample 
+- `spatSample<SpatRaster>` and `spatSample<SpatExtent>` argument gain argument "exact=FALSE" to request the exact (but perhaps less regular) sample size for a regular sample. Currently pnly for planar crs.
 
 ## new